requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
pyaes==1.6.1
pyasn1==0.4.8
python-dotenv==1.0.0
python-telegram-bot==20.2
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
sniffio==1.3.0
telegram==0.0.1
Telethon==1.27.0
threaded==4.1.0
urllib3==1.26.15
